Add support for Spark Dynamic Resource allocation.
This change adds the Spark Dynamic Resource allocation support enabled
by default. It triggers a failure when the option is used in standalone mode
since it is only availabile in YARN mode. In this way the feature will be
automatically enabled without any other puppet change in the main puppet repo
(standalone mode is supported by this module but we are not currently using it).

Bug: T101343
Change-Id: Ice2337f86ed1758af435450b00132e79312d33b2