oozie: override the oozie-setup script
The oozie-setup script uses 'su -s /bin/bash -c etc..'
to copy libraries over to HDFS. If the user running the script
cannot become root then the script stops asking for a password.
Tried on the testing cluster and worked as expected.

Before the introduction of cdh::exec this script was executed
by root, then we changed it to avoid using unnecessary privileges.
In the future it will also be easier to assign a kerberos identity to
'oozie' rather than 'root'.

Bug: T218343
Change-Id: I7b311087cb9f27d0d6459c594363d102543ce392