Add the dfs.namenode.handler.count HDFS option
From https://community.hortonworks.com/articles/43839/scaling-the-hdfs-namenode-part-2.html:
"The Hadoop RPC server consists of a single RPC queue per port and multiple handler (worker)
threads that dequeue and process requests. If the number of handlers is insufficient,
then the RPC queue starts building up and eventually overflows.
You may start seeing task failures and eventually job failures and unhappy users.
It is recommended that the RPC handler count be set to 20 * log2(Cluster Size) with an upper limit of 200."

The current hadoop default is 10.

Bug: T220702
Change-Id: I183b1af85ef6e9b65b772d8751ac9a160b3c58bd